Year 2021 Day 9

+++ src/util/grid.rs
@@ -2,8 +2,8 @@ use crate::util::point::*;
 
 #[derive(Clone)]
 pub struct Grid {
-    pub width: usize,
-    pub height: usize,
+    pub width: i32,
+    pub height: i32,
     pub bytes: Vec<u8>,
 }
 
@@ -13,9 +13,9 @@ impl Grid {
             .lines()
             .map(|line| line.as_bytes())
             .collect();
-        let width = raw[0].len();
-        let height = raw.len();
-        let mut bytes = Vec::with_capacity(width * height);
+        let width = raw[0].len() as i32;
+        let height = raw.len() as i32;
+        let mut bytes = Vec::with_capacity((width * height) as usize);
         raw.iter().for_each(|slice| bytes.extend_from_slice(slice));
         Grid { width, height, bytes }
     }
@@ -24,7 +24,7 @@ impl Grid {
         Grid {
             width: self.width,
             height: self.height,
-            bytes: vec![0; self.width * self.height],
+            bytes: vec![0; (self.width * self.height) as usize],
         }
     }
 
@@ -36,17 +36,17 @@ impl Grid {
     }
 
     pub fn get(&self, point: Point) -> u8 {
-        self.bytes[self.width * (point.y as usize) + (point.x as usize)]
+        self.bytes[(self.width * point.y + point.x) as usize]
     }
 
     pub fn set(&mut self, point: Point, value: u8) {
-        self.bytes[self.width * (point.y as usize) + (point.x as usize)] = value;
+        self.bytes[(self.width * point.y + point.x) as usize] = value;
     }
 
     pub fn find(&self, needle: u8) -> Option<Point> {
         let to_point = |index| {
-            let x = (index % self.width) as i32;
-            let y = (index / self.width) as i32;
+            let x = (index as i32) % self.width;
+            let y = (index as i32) / self.width;
             Point { x, y }
         };
         self.bytes.iter().position(|&h| h == needle).map(to_point)
